app/requirements.txt
# Core dependencies
pymupdf==1.23.8
spacy==3.7.2
python-docx==1.1.0
ocrmypdf==15.4.3
pandas==2.1.3
numpy==1.24.3

# NLP and ML
scikit-learn==1.3.0
sentence-transformers==2.2.2
networkx==3.1

# Web Interface
streamlit==1.28.1
plotly==5.17.0

# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0

# Reporting
reportlab==4.0.4
matplotlib==3.8.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3.post1

# Development
flake8==6.1.0
pytest==7.4.3
black==23.11.0
pytest-cov==4.1.0

# Optional for enhanced features
# Uncomment if needed:
# pdf2image==1.16.3
# pillow==10.1.0
# opencv-python==4.8.1.78
# transformers==4.35.2
# torch==2.1.0